refactor: use u16 over u32

+++ src/fixed/lifeless.rs
@@ -6,8 +6,8 @@ pub type Identifier = String;
 #[cfg_attr(test, derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase")]
 pub struct Lifeless {
-    pub hitpoints_armor: u32,
-    pub hitpoints_structure: u32,
+    pub hitpoints_armor: u16,
+    pub hitpoints_structure: u16,
     // TODO: mineable resources
     // TODO: lootable resources
     // TODO: hackable resources
